add EntityContext to replace EntityQueryable
