+++ Wodsoft.ComBoost.EntityFramework/Data/Entity/EntityExtensions.cs
@@ -6,8 +6,16 @@ using System.Threading.Tasks;
 
 namespace System.Data.Entity
 {
+    /// <summary>
+    /// Extensions of entity.
+    /// </summary>
     public static class EntityExtensions
     {
+        /// <summary>
+        /// Get database context from a context builder.
+        /// </summary>
+        /// <param name="builder">Entity context builder.</param>
+        /// <returns></returns>
         public static DbContext GetContext(IEntityContextBuilder builder)
         {
             if (builder == null)
